+++ Tests/AdvancedMappingCrud.RichDomain.ServiceModel.Tests/AdvancedMappingCrud.RichDomain.ServiceModel.Tests.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -95,8 +95,7 @@ namespace AdvancedMappingCrud.RichDomain.ServiceModel.Tests.Infrastructure.Persi
             {
                 var domainEventEntity = ChangeTracker
                     .Entries<IHasDomainEvent>()
-                    .Select(x => x.Entity.DomainEvents)
-                    .SelectMany(x => x)
+                    .SelectMany(x => x.Entity.DomainEvents)
                     .FirstOrDefault(domainEvent => !domainEvent.IsPublished);
 
                 if (domainEventEntity is null)