+++ Tests/AdvancedMappingCrud.RichDomain.Tests/AdvancedMappingCrud.RichDomain.Tests.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -96,8 +96,7 @@ namespace AdvancedMappingCrud.RichDomain.Tests.Infrastructure.Persistence
             {
                 var domainEventEntity = ChangeTracker
                     .Entries<IHasDomainEvent>()
-                    .Select(x => x.Entity.DomainEvents)
-                    .SelectMany(x => x)
+                    .SelectMany(x => x.Entity.DomainEvents)
                     .FirstOrDefault(domainEvent => !domainEvent.IsPublished);
 
                 if (domainEventEntity is null)