+++ Tests/AzureFunctions.AzureEventGrid/AzureFunctions.AzureEventGrid.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -61,8 +61,7 @@ namespace AzureFunctions.AzureEventGrid.Infrastructure.Persistence
             {
                 var domainEventEntity = ChangeTracker
                     .Entries<IHasDomainEvent>()
-                    .Select(x => x.Entity.DomainEvents)
-                    .SelectMany(x => x)
+                    .SelectMany(x => x.Entity.DomainEvents)
                     .FirstOrDefault(domainEvent => !domainEvent.IsPublished);
 
                 if (domainEventEntity is null)