+++ Tests/CleanArchitecture.Comprehensive/CleanArchitecture.Comprehensive.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -192,8 +192,7 @@ namespace CleanArchitecture.Comprehensive.Infrastructure.Persistence
             {
                 var domainEventEntity = ChangeTracker
                     .Entries<IHasDomainEvent>()
-                    .Select(x => x.Entity.DomainEvents)
-                    .SelectMany(x => x)
+                    .SelectMany(x => x.Entity.DomainEvents)
                     .FirstOrDefault(domainEvent => !domainEvent.IsPublished);
 
                 if (domainEventEntity is null)