+++ Tests/EntityFrameworkCore.MultiDbContext.DbContextInterface/EntityFrameworkCore.MultiDbContext.DbContextInterface.Infrastructure/Persistence/ConnStrDbContext.cs
@@ -66,8 +66,7 @@ namespace EntityFrameworkCore.MultiDbContext.DbContextInterface.Infrastructure.P
             {
                 var domainEventEntity = ChangeTracker
                     .Entries<IHasDomainEvent>()
-                    .Select(x => x.Entity.DomainEvents)
-                    .SelectMany(x => x)
+                    .SelectMany(x => x.Entity.DomainEvents)
                     .FirstOrDefault(domainEvent => !domainEvent.IsPublished);
 
                 if (domainEventEntity is null)