+++ Tests/FastEndpointsTest/FastEndpointsTest.Infrastructure/Persistence/ApplicationDbContext.cs
@@ -115,8 +115,7 @@ namespace FastEndpointsTest.Infrastructure.Persistence
             {
                 var domainEventEntity = ChangeTracker
                     .Entries<IHasDomainEvent>()
-                    .Select(x => x.Entity.DomainEvents)
-                    .SelectMany(x => x)
+                    .SelectMany(x => x.Entity.DomainEvents)
                     .FirstOrDefault(domainEvent => !domainEvent.IsPublished);
 
                 if (domainEventEntity is null)